requirements.txt
Django==4.1.3
django-environ
psycopg2-binary
djangorestframework==3.14.0
social-auth-app-django==5.0.0
django-filter~=22.1
pydantic==1.10.2
requests==2.28.1
pytest-django==4.5.2
pytest_factoryboy==2.5.0